Merge controls into master

+++ src/wrecker/src/drivers/rmMotor.cpp
@@ -8,7 +8,7 @@ short rmMotor::getRpm() {
     return rpm;
 }
 
-short rmMotor::getAngle() {
+float rmMotor::getAngle() {
     return angle;
 }
